refactored subjectController, services and dtos related

+++ src/main/java/com/secondcommit/forum/controllers/SubjectController.java
@@ -1,6 +1,7 @@
 package com.secondcommit.forum.controllers;
 
 import com.secondcommit.forum.dto.SubjectDto;
+import com.secondcommit.forum.repositories.SubjectRepository;
 import com.secondcommit.forum.security.payload.MessageResponse;
 import com.secondcommit.forum.services.subject.SubjectServiceImpl;
 import io.swagger.annotations.ApiOperation;
@@ -17,15 +18,17 @@ import org.springframework.web.bind.annotation.*;
 public class SubjectController {
 
     private SubjectServiceImpl subjectService;
+    private SubjectRepository subjectRepository;
 
-    public SubjectController(SubjectServiceImpl subjectService) {
+    public SubjectController(SubjectServiceImpl subjectService, SubjectRepository subjectRepository) {
         this.subjectService = subjectService;
+        this.subjectRepository = subjectRepository;
     }
 
     /**
-     * Creates a new subject from subjectDto
+     * Creates a new subject from subjectDto. ADMIN only
      * @param subjectDto
-     * @return ResponseEntity<Subject>
+     * @return ResponseEntity (ok: Subject, bad request: messageResponse)
      */
     @PreAuthorize("hasAuthority('ADMIN')")
     @PostMapping("/")
@@ -39,7 +42,7 @@ public class SubjectController {
     }
 
     /**
-     * Method to get a specific subject
+     * Method to get all data in a specific subject
      * @param id
      * @return ResponseEntity (ok: subject, bad request: messageResponse)
      */
@@ -47,6 +50,11 @@ public class SubjectController {
     @GetMapping("/{id}")
     @ApiOperation("Gets all subject data")
     public ResponseEntity<?> getSubject(@PathVariable Long id){
+
+        //Validates the subject
+        if (!subjectRepository.existsById(id))
+            return ResponseEntity.badRequest().body(new MessageResponse("Wrong id"));
+
         return subjectService.getSubject(id);
     }
 
@@ -57,39 +65,48 @@ public class SubjectController {
      */
     @PreAuthorize("hasAuthority('USER')")
     @GetMapping("/")
-    @ApiOperation("Gets all subject data")
+    @ApiOperation("Gets all subjects which the user has access to")
     public ResponseEntity<?> getSubjectsAllowed(@CurrentSecurityContext(expression="authentication?.name") String username){
         return subjectService.getSubjectsAllowed(username);
     }
 
     /**
-     * Method that updates the subject
+     * Method that updates the subject. ADMIN only
      * @param id
      * @param subjectDto
-     * @return ResponseEntity<Subject>
+     * @return ResponseEntity (ok: Subject, bad request: messageResponse)
      */
     @PreAuthorize("hasAuthority('ADMIN')")
     @PutMapping("/{id}")
     @ApiOperation("Updates subject data")
     public ResponseEntity<?> updateSubject(@PathVariable Long id, @RequestBody SubjectDto subjectDto){
 
+        //Validates DTO
         if (subjectDto.getName() == null && subjectDto.getModules() == null)
             return ResponseEntity.badRequest().body(new MessageResponse("Missing parameters"));
+
+        //Validates ID
+        if (!subjectRepository.existsById(id))
+            return ResponseEntity.badRequest().body(new MessageResponse("Wrong id"));
 
         return subjectService.updateSubject(id, subjectDto);
     }
 
     /**
-     * Deletes subject
+     * Method to delete subject. ADMIN only
      * @param id
-     * @return ResponseEntity
+     * @return ResponseEntity (messageResponse)
      */
     @PreAuthorize("hasAuthority('ADMIN')")
     @DeleteMapping("/{id}")
     @ApiOperation("Deletes subject")
     public ResponseEntity<?> deleteSubject(@PathVariable Long id){
+
+        //Validates the id
+        if (!subjectRepository.existsById(id))
+            return ResponseEntity.badRequest().body(new MessageResponse("Wrong id"));
+
         return subjectService.deleteSubject(id);
     }
 
-    //TODO: Create DTO for the module and update the subject response dto. It's sending too much information back
 }
